Finishing Login/Register handling
